refactor(utils): 重构工具函数模块
- 移除未使用的导入模块
- 将服务器相关功能移至独立的serverManager模块
- 优化服务器状态切换逻辑
- 改进服务器配置加载方式
- 更新模块导入路径和依赖关系
- 优化列表推导式为显式循环以提高可读性
